+++ domain/src/main/java/com/example/domain/entity/responses/CreationAnEventResponseEntity.kt
@@ -29,7 +29,7 @@ data class CreationAnEventResponseEntityData(
 class CreationAnEventResponseEntityForms
 
 data class CreationAnEventResponseEntityPlace(
-    val lat: Int,
-    val lon: Int,
+    val lat: Double,
+    val lon: Double,
     val place_name: String
 )
